serverless/pubsub/template_pubsub/requirements.txt
google-cloud-pubsub==2.1.0
pytz